+++ vtm/src/org/oscim/backend/XMLReaderAdapter.java
@@ -1,5 +1,6 @@
 /*
  * Copyright 2013 Hannes Janetzek
+ * Copyright 2016 devemux86
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
  *
@@ -28,24 +29,16 @@ import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParserFactory;
 
 /**
- * The Class XMLReaderAdapter. Used by RenderTheme. Note: GWT Backend provides
- * its own implementation.
+ * The class XMLReaderAdapter is used by RenderTheme.
+ * Note: GWT Backend provides its own implementation.
  */
 public class XMLReaderAdapter {
-    public void parse(DefaultHandler handler, InputStream is) throws IOException {
-
+    public void parse(DefaultHandler handler, InputStream is) throws ParserConfigurationException, SAXException, IOException {
         SAXParserFactory factory = SAXParserFactory.newInstance();
         factory.setNamespaceAware(true);
 
-        try {
-            XMLReader xmlReader = factory.newSAXParser().getXMLReader();
-            xmlReader.setContentHandler(handler);
-            xmlReader.parse(new InputSource(is));
-
-        } catch (SAXException e) {
-            e.printStackTrace();
-        } catch (ParserConfigurationException e) {
-            e.printStackTrace();
-        }
+        XMLReader xmlReader = factory.newSAXParser().getXMLReader();
+        xmlReader.setContentHandler(handler);
+        xmlReader.parse(new InputSource(is));
     }
 }