+++ vtm/src/org/oscim/theme/ThemeLoader.java
@@ -19,52 +19,21 @@ package org.oscim.theme;
 
 import org.oscim.backend.CanvasAdapter;
 import org.oscim.theme.IRenderTheme.ThemeException;
-import org.oscim.utils.IOUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.FileNotFoundException;
-import java.io.InputStream;
 
 public class ThemeLoader {
-    static final Logger log = LoggerFactory.getLogger(ThemeLoader.class);
 
-    /**
-     * Load theme from XML file.
-     *
-     * @throws FileNotFoundException
-     * @throws ThemeException
-     */
-    public static IRenderTheme load(String renderThemePath) throws ThemeException,
-            FileNotFoundException {
+    public static IRenderTheme load(String renderThemePath) throws ThemeException {
         return load(new ExternalRenderTheme(renderThemePath));
     }
 
+    public static IRenderTheme load(String renderThemePath, XmlRenderThemeMenuCallback menuCallback) throws ThemeException {
+        return load(new ExternalRenderTheme(renderThemePath, menuCallback));
+    }
+
     public static IRenderTheme load(ThemeFile theme) throws ThemeException {
-
-        try {
-            InputStream is = theme.getRenderThemeAsStream();
-            return load(theme.getRelativePathPrefix(), is);
-        } catch (FileNotFoundException e) {
-            log.error(e.getMessage());
-        }
-
-        return null;
-    }
-
-    public static IRenderTheme load(InputStream inputStream) throws ThemeException {
-        return load("", inputStream);
-    }
-
-    public static IRenderTheme load(String relativePathPrefix, InputStream inputStream) throws ThemeException {
-
-        try {
-            IRenderTheme t = XmlThemeBuilder.read(relativePathPrefix, inputStream);
-            if (t != null)
-                t.scaleTextSize(CanvasAdapter.textScale + (CanvasAdapter.dpi / 240 - 1) * 0.5f);
-            return t;
-        } finally {
-            IOUtils.closeQuietly(inputStream);
-        }
+        IRenderTheme t = XmlThemeBuilder.read(theme);
+        if (t != null)
+            t.scaleTextSize(CanvasAdapter.textScale + (CanvasAdapter.dpi / 240 - 1) * 0.5f);
+        return t;
     }
 }