+++ vtm/src/org/oscim/theme/XmlRenderThemeMenuCallback.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2014 Ludwig M Brinckmann
+ * Copyright 2016 devemux86
+ *
+ * This program is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU Lesser General Public License as published by the Free Software
+ * Foundation, either version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+ * PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License along with
+ * this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.oscim.theme;
+
+import java.util.Set;
+
+/**
+ * Interface callback in render theme to create a settings menu on the fly.
+ */
+public interface XmlRenderThemeMenuCallback {
+    /**
+     * Called when the style menu section of the xml file has been parsed.
+     */
+    Set<String> getCategories(XmlRenderThemeStyleMenu style);
+}